chore : split axios config and util file
