+++ package.json
@@ -38,5 +38,11 @@
     "typescript-eslint": "^8.22.0",
     "vite": "^6.1.0",
     "vite-tsconfig-paths": "^5.1.4"
+  },
+  "pnpm": {
+    "ignoredBuiltDependencies": [
+      "@swc/core",
+      "esbuild"
+    ]
   }
 }
