+++ src/apis/auth/mutations/useAuthJoinMutation..ts
@@ -1,8 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
-
 import { AuthJoinRequest } from "@apis/auth/models/AuthJoin";
 import { RestErrorResponse, RestResponse } from "@models/io/RestResponse";
-
 import rest from "@utils/rest";
 
 const useAuthJoinMutation = () => {
